refactor(resources/calculator): unified folder
